game: implement ball taking and also refactor code
Bird movement is handled by a callback that is handed to the bird.
This allows for DRY code when considering keybindings.
The bird now has code for movement but it is injected by the view.
This is different to the movement of the ball, but only because the ball
does not respond to keypresses (as of yet). Should consider changing in
future if that also changes.

Mentions #3.